Move all styling local to components
Fixes #13
