+++ src/App.js
@@ -2,8 +2,8 @@ import React, { Component } from 'react';
 import './App.css';
 
 // Components
-import Button from './components/Button';
-import Card from './components/Card';
+import Button from './components/Button/Button';
+import Card from './components/Card/Card';
 
 class App extends Component {
