+++ src/components/Button/Button.js
@@ -1,9 +1,10 @@
 import React from 'react';
+import './Button.css';
 
 class Button extends React.Component {
   render() {
     return (
-      <button className={"p-button--" + this.props.type}>{this.props.value}</button>
+      <button className={`p-button--${this.props.type}`}>{this.props.value}</button>
     )
   }
 }
