+++ src/components/Button/Button.scss
@@ -1,0 +1,3 @@
+@import 'vanilla-framework/scss/base_button';
+@import 'vanilla-framework/scss/patterns_buttons';
+@include vf-p-buttons;
