+++ src/components/Card/Card.scss
@@ -1,0 +1,2 @@
+@import 'vanilla-framework/scss/patterns_card';
+@include vf-p-card;
